+++ packages/ragbits-core/src/ragbits/core/embeddings/base.py
@@ -1,5 +1,4 @@
 from abc import ABC, abstractmethod
-from enum import Enum
 from typing import ClassVar, TypeVar
 
 from ragbits.core import embeddings
@@ -7,21 +6,6 @@ from ragbits.core.options import Options
 from ragbits.core.utils.config_handling import ConfigurableComponent
 
 EmbedderOptionsT = TypeVar("EmbedderOptionsT", bound=Options)
-
-
-class EmbeddingType(Enum):
-    """
-    Indicates the type of embedding in regard to what kind of data has been embedded.
-
-    Used to specify the embedding type for a given element. Unlike `Element` type,
-    which categorizes the element itself, `EmbeddingType` determines how the
-    element's data is represented. For example, an image element can support
-    multiple embedding types, such as a description, OCR output, or raw bytes,
-    allowing for the creation of different embeddings for the same element.
-    """
-
-    TEXT = "text"
-    IMAGE = "image"
 
 
 class Embedder(ConfigurableComponent[EmbedderOptionsT], ABC):