+++ packages/ragbits-document-search/tests/unit/testprojects/project_with_instance_factory/factories.py
@@ -4,12 +4,15 @@ from ragbits.core.embeddings.noop import NoopEmbedder
 from ragbits.core.vector_stores.in_memory import InMemoryVectorStore, VectorStoreOptions
 from ragbits.document_search import DocumentSearch
 from ragbits.document_search.documents.document import DocumentMeta
+from ragbits.document_search.retrieval.rerankers.base import RerankerOptions
+from ragbits.document_search.retrieval.rerankers.noop import NoopReranker
 
 
 def create_document_search_instance_223():
     vector_store_options = VectorStoreOptions(k=223)
     document_search = DocumentSearch(
-        embedder=NoopEmbedder(), vector_store=InMemoryVectorStore(default_options=vector_store_options)
+        reranker=NoopReranker(default_options=RerankerOptions(top_n=223)),
+        vector_store=InMemoryVectorStore(embedder=NoopEmbedder(), default_options=vector_store_options),
     )
     return document_search
 
@@ -17,7 +20,8 @@ def create_document_search_instance_223():
 def create_document_search_instance_825():
     vector_store_options = VectorStoreOptions(k=825)
     document_search = DocumentSearch(
-        embedder=NoopEmbedder(), vector_store=InMemoryVectorStore(default_options=vector_store_options)
+        reranker=NoopReranker(default_options=RerankerOptions(top_n=825)),
+        vector_store=InMemoryVectorStore(embedder=NoopEmbedder(), default_options=vector_store_options),
     )
     return document_search
 
@@ -32,6 +36,6 @@ async def _add_example_documents(document_search: DocumentSearch) -> None:
 
 
 def create_document_search_instance_with_documents():
-    document_search = DocumentSearch(embedder=NoopEmbedder(), vector_store=InMemoryVectorStore())
+    document_search = DocumentSearch(vector_store=InMemoryVectorStore(embedder=NoopEmbedder()))
     asyncio.run(_add_example_documents(document_search))
     return document_search